+++ frontend/src/views/TodoList/TodoList.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { usePageTitle } from 'providers/PageTitleProvider';
 import { TodoTaskList } from 'components/TodoTaskList/TodoTaskList';
-import { todoTaskListMock } from 'mock-data/TodoTaskList/mock-1';
+import { todoTaskListMock } from 'mock-data/TodoTaskList/todoTaskListMock';
 
 export const TodoList = () => {
   const { setTitle } = usePageTitle();
